db_manager/requirements.txt
psycopg2-binary
time